revert index and output files to original state
