--- output.js
@@ -1,182 +1,0 @@
-export const legacyOutput = {
-  "0": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "1": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "2": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "3": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "4": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "5": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "6": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "7": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "8": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "9": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "10": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "11": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "12": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "13": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "14": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "15": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "16": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "17": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "18": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "19": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "20": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "21": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "22": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "23": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "24": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "25": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "26": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "27": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "28": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-  "29": [
-    { name: "Doliprane", expiresIn: -10, benefit: 0 },
-    { name: "Herbal Tea", expiresIn: -20, benefit: 50 },
-    { name: "Fervex", expiresIn: -25, benefit: 0 },
-    { name: "Magic Pill", expiresIn: 15, benefit: 40 },
-  ],
-};
